UI fixed and show mode state changes finished
