+++ src/lib/Nav.svelte
@@ -1,19 +1,21 @@
 <script>
 	import { rentalType } from "$lib/stores"
 
-  let dateObj = new Date();
+    export let mainHeader = "Welcome to the Registry"; // Default value if not provided
+
+    let dateObj = new Date();
 	/** @type {Intl.DateTimeFormatOptions} */
-	let options = {
-		weekday: 'long',
-		year: 'numeric',
-		month: 'long',
-		day: 'numeric'
-	};
-	let currentDate = new Intl.DateTimeFormat('en-US', options)
-		.format(dateObj)
-		.split(',')
-		.slice(0, 2)
-		.join(', ');
+    let options = {
+        weekday: 'long',
+        year: 'numeric',
+        month: 'long',
+        day: 'numeric'
+    };
+    let currentDate = new Intl.DateTimeFormat('en-US', options)
+        .format(dateObj)
+        .split(',')
+        .slice(0, 2)
+        .join(', ');
 
     let currentRentalType;
     rentalType.subscribe(value => {
@@ -24,9 +26,9 @@
 <div id="logo-container">
   <img src="/registry_logo.png" alt="Theatre Logo" id="theatre-logo">
 </div>
-<h1>Welcome to the Registry.</h1>
+<h1>{mainHeader}</h1> <!-- Use the mainHeader prop here -->
 <div id="header-info">
-	<p>{currentRentalType}</p>
+    <p>{currentRentalType}</p>
     <p>{currentDate}</p>
 </div>
 
@@ -36,5 +38,4 @@
     height: auto;
     object-fit: contain;
   }
-
 </style>
